+++ core/test/org/javarosa/core/model/test/FormDefTest.java
@@ -17,7 +17,7 @@ import org.javarosa.core.model.data.StringData;
 import org.javarosa.core.model.instance.FormInstance;
 import org.javarosa.core.model.instance.TreeElement;
 import org.javarosa.core.model.instance.TreeReference;
-import org.javarosa.core.model.instance.DummyInstanceInitializationFactory;
+import org.javarosa.core.model.instance.test.DummyInstanceInitializationFactory;
 import org.javarosa.core.services.PrototypeManager;
 import org.javarosa.core.test.FormParseInit;
 import org.javarosa.core.util.externalizable.ExtUtil;
