+++ core/test/org/javarosa/xpath/expr/test/XPathPathExprTest.java
@@ -124,6 +124,9 @@ public class XPathPathExprTest extends TestCase {
                 fd.getInstance(), null, new XPathTypeMismatchException());
     }
 
+    /**
+     * Test nested predicates that have relative and absolute references.
+     */
     private void testNestedPreds() {
         FormParseInit fpi = new FormParseInit("/test_nested_preds_with_rel_refs.xml");
         FormDef fd = fpi.getFormDef();
@@ -139,15 +142,16 @@ public class XPathPathExprTest extends TestCase {
         testEval("join(' ', instance('groups')/root/groups/group/@id)",
                 groupsInstance, ec, "inc dwa");
 
-        testEval("count(instance('groups')/root/groups[position() = 1]/team[@id = 'mobile'])", groupsInstance, ec, 1.0);
+        testEval("count(instance('groups')/root/groups[position() = 1]/team[@id = 'mobile'])",
+                groupsInstance, ec, 1.0);
 
         // find 'group' elements that have a 'team' sibling with id = mobile;
-        // should be 'inc'
-        testEval("instance('groups')/root/groups/group[count(../team[@id = 'mobile']) > 0]/@id", groupsInstance, ec, "inc");
-        testEval("count(instance('groups')/root/groups/group[count(../team[@id = 'mobile']) > 0])", groupsInstance, ec, 1.0);
+        testEval("instance('groups')/root/groups/group[count(../team[@id = 'mobile']) > 0]/@id",
+                groupsInstance, ec, "inc");
 
-        testEval("count(instance('groups')/root/groups/group/group_data/data) > 0", groupsInstance, ec, true);
-        testEval("count(instance('groups')/root/groups/group[count(../team[@id = 'mobile']) > 0]) = 1", groupsInstance, ec, true);
+        testEval("count(instance('groups')/root/groups/group[count(../team[@id = 'mobile']) > 0]) = 1",
+                groupsInstance, ec, true);
+
         testEval("if(count(instance('groups')/root/groups/group/group_data/data) > 0 and count(instance('groups')/root/groups/group[count(../team[@id = 'mobile']) > 0]) = 1, instance('groups')/root/groups/group[count(../team[@id = 'mobile']) > 0]/@id, '')",
                 groupsInstance, ec, "inc");
